% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{as.data.frame.buscador}
\alias{as.data.frame.buscador}
\title{generic for converting `buscador` search object to a dataframe}
\usage{
\method{as.data.frame}{buscador}(x, ...)
}
\arguments{
\item{x}{`buscador` object to coerce}

\item{...}{parameters for other functions}
}
\description{
generic for converting `buscador` search object to a dataframe
}
