% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{as.data.frame.busco}
\alias{as.data.frame.busco}
\title{generic for converting `busco` searh object to a dataframe}
\usage{
\method{as.data.frame}{busco}(x, ...)
}
\arguments{
\item{x}{`busco` object to coerce}

\item{...}{parameters for other functions}
}
\description{
generic for converting `busco` searh object to a dataframe
}
