% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{as.drawProteins}
\alias{as.drawProteins}
\title{get a dataframe of receptor type data in `drawProteins` format
for pretty drawing. From a `buscador` search object from `buscar()`}
\usage{
as.drawProteins(b, which = "lrr_rp")
}
\arguments{
\item{b}{buscador search object returned from `buscar`}

\item{which}{the receptor type to return, one of 'lrr_rp', 'lrr_rk', 'non_lrr_rp', 'non_lrr_rk', 'lrr_rp_ecto'}
}
\value{
dataframe
}
\description{
get a dataframe of receptor type data in `drawProteins` format
for pretty drawing. From a `buscador` search object from `buscar()`
}
