% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{as_drawProteins}
\alias{as_drawProteins}
\title{get a dataframe of receptor type data in `drawProteins` format
for pretty drawing. From a `buscador` search object from `buscar()`}
\usage{
as_drawProteins(b, which = "LRR-RP")
}
\arguments{
\item{b}{buscador search object returned from `buscar`}

\item{which}{the receptor type to return, one of}
}
\value{
dataframe
}
\description{
get a dataframe of receptor type data in `drawProteins` format
for pretty drawing. From a `buscador` search object from `buscar()`
}
