% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{as_tibble}
\alias{as_tibble}
\title{converts `buscador` search object to a tibble}
\usage{
as_tibble(b)
}
\arguments{
\item{b}{`buscador` object to coerce}
}
\description{
converts `buscador` search object to a tibble
}
