% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{blast_results}
\alias{blast_results}
\title{get the dataframe of ALL blast results from a `buscar` search object from `buscar()`}
\usage{
blast_results(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of ALL blast results from a `buscar` search object from `buscar()`
}
