% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{classification_matrix}
\alias{classification_matrix}
\title{get the computed classification matrix for the proteins}
\usage{
classification_matrix(b)
}
\arguments{
\item{b}{buscador object}
}
\value{
dataframe
}
\description{
get the computed classification matrix for the proteins
}
