% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{classify_protein}
\alias{classify_protein}
\title{join the classification from the def matrix to the attributes determined
by the blast hit}
\usage{
classify_protein(b)
}
\arguments{
\item{b}{buscador pbkect}
}
\description{
join the classification from the def matrix to the attributes determined
by the blast hit
}
