% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{completed}
\alias{completed}
\title{check the status of the buscador object}
\usage{
completed(busc)
}
\arguments{
\item{busc}{a buscador object}
}
\description{
are all the searches done and the results compiled
}
