% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{condense}
\alias{condense}
\title{turn the tidy long format dataframe into a wider sequence per line dataframe}
\usage{
condense(df)
}
\arguments{
\item{df}{dataframe to condense}
}
\value{
dataframe
}
\description{
turn the tidy long format dataframe into a wider sequence per line dataframe
}
