% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{condense_blasts}
\alias{condense_blasts}
\title{turn the tidy long format blast dataframe into a wider sequence per line dataframe}
\usage{
condense_blasts(b)
}
\arguments{
\item{b}{buscador object}
}
\value{
dataframe
}
\description{
turn the tidy long format blast dataframe into a wider sequence per line dataframe
}
