% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{condense_pfams}
\alias{condense_pfams}
\title{turn the tidy long format pfams dataframe into a wider sequence per line dataframe}
\usage{
condense_pfams(b)
}
\arguments{
\item{b}{buscador objects}
}
\value{
dataframe
}
\description{
turn the tidy long format pfams dataframe into a wider sequence per line dataframe
}
