% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{deeptmhmm_results}
\alias{deeptmhmm_results}
\title{get the dataframe of passing deeptmhmm results from a `buscar` search object from `buscar()`}
\usage{
deeptmhmm_results(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of passing deeptmhmm results from a `buscar` search object from `buscar()`
}
