% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defs.R
\name{definiciones}
\alias{definiciones}
\title{report the definitions used for each receptor type}
\usage{
definiciones(which = "all")
}
\arguments{
\item{which}{the type to show a definition for; one of
'all', 'lrr_rp', 'lrr_rk', 'non_lrr_rp', 'non_lrr_rk', 'lrr_rp_ecto'}
}
\value{
character vector
}
\description{
report the definitions used for each receptor type
}
