% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{dibujar}
\alias{dibujar}
\title{draw each found proteins of a given receptor type.}
\usage{
dibujar(b, which = "LRR-RP", label_domains = FALSE)
}
\arguments{
\item{b}{`buscador` search object returned from `buscar()`}

\item{which}{the receptor type to return, one of}

\item{label_domains}{write a label on the domain}
}
\value{
ggplot2
}
\description{
draw each found proteins of a given receptor type.
}
