% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{do_searches}
\alias{do_searches}
\title{run searches over the internet and local BLAST.}
\usage{
do_searches(busc)
}
\arguments{
\item{busc}{buscador object from `buscar`}
}
\value{
search populated buscador object
}
\description{
Runs searches for proteins at Phobius, PFAM and local BLAST.
Reformat returned results from searches and apply basic quality filters
}
\details{
if the search at PFAM does not complete the buscador$pfam_progress object is
populated with submission IDs for a later restart run and the restart_file written.
In later runs the PFAM server is queried to collect all completed runs. This process should
be repeated until all the queries are complete. Once all queries are complete the buscador$pfam object gets the
dataframe of all PFAM results and is marked with class "complete"
}
