% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecto_dom_seqs}
\alias{ecto_dom_seqs}
\title{ecto domain sequences from Arabidopsis for BLAST}
\format{
Bioconductor DNAStringSet
}
\usage{
ecto_dom_seqs
}
\description{
ecto domain sequences from Arabidopsis for BLAST
}
\keyword{datasets}
