% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{ecto_results}
\alias{ecto_results}
\title{get the dataframe of Ecto domain results from a `busco` search object from `buscar()`}
\usage{
ecto_results(b)
}
\arguments{
\item{b}{busco search object returned from `busco`}
}
\value{
dataframe
}
\description{
get the dataframe of Ecto domain results from a `busco` search object from `buscar()`
}
