% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{find_tm_proteins}
\alias{find_tm_proteins}
\title{get vector of seq_names with 1 tm domain and 1 signal peptide}
\usage{
find_tm_proteins(dtm)
}
\arguments{
\item{dtm}{parsed deeptmhmm result from `process_raw_deeptmhmm`}
}
\value{
character vector
}
\description{
get vector of seq_names with 1 tm domain and 1 signal peptide
}
