% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecto.R
\name{get_ecto}
\alias{get_ecto}
\title{Run BLASTP using protein file as database and ectodomains as query}
\usage{
get_ecto(file, progress = FALSE)
}
\arguments{
\item{file}{protein fasta file}

\item{progress}{show a progress bar}
}
\description{
Run BLASTP using protein file as database and ectodomains as query
}
