% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfam.R
\name{get_pfam}
\alias{get_pfam}
\title{Query PFAM web server.}
\source{
\url{https://www.ebi.ac.uk/Tools/common/tools/help/index.html}
}
\usage{
get_pfam(file, email, progress = FALSE, eval = 20)
}
\arguments{
\item{file}{A character file path of FASTA protein amino acid sequences a}

\item{email}{Character, a valid user email address, needed for PFAMscan use}

\item{progress}{Boolean, whether to show the progress bar, at default set to
FALSE.}

\item{eval}{PFAM hit eval cutoff}
}
\value{
A data frame with columns:
\describe{
\item{seq_name}{Character, name of the submitted sequence.}
\item{hit}{Character, the name of a PFAM hit.}
\item{acc}{Character, the accession of a PFAM hit.}
\item{eval}{Double, E-value of the hit.}
\item{type}{Character, PFAM type of the hit}
\item{seq_from}{Integer, start of match on sequence}
\item{seq_to}{Integer, end of match on sequence}
\item{hit_from}{Integer, start of match on hit}
\item{hit_to}{Integer, end of match on hit}
}
}
\description{
Query PFAM web server.
}
\references{
Madeira F, Park YM, Lee J, et al. The EMBL-EBI search and sequence analysis
tools APIs in 2019. Nucleic Acids Research. 2019 Jul;47(W1):W636-W641.
\doi{10.1093/nar/gkz268}.
}
