% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{groups}
\alias{groups}
\title{get the dataframe of classification results from a `buscar` search object from `buscar()`}
\usage{
groups(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of classification results from a `buscar` search object from `buscar()`
}
