% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{has_at_lrr_blast}
\alias{has_at_lrr_blast}
\title{work out whether a sequence has lrr blast according to criteria}
\usage{
has_at_lrr_blast(sn, b)
}
\arguments{
\item{sn}{vector of sequence names with TM and SP}

\item{b}{buscador object}
}
\description{
work out whether a sequence has lrr blast according to criteria
}
