% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{has_kinase_pfam}
\alias{has_kinase_pfam}
\title{work out whether a sequence has kinase pfams according to criteria}
\usage{
has_kinase_pfam(sn, b)
}
\arguments{
\item{sn}{vector of sequence names with TM and SP}

\item{b}{buscador object}
}
\description{
work out whether a sequence has kinase pfams according to criteria
}
