% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{hmmer_results}
\alias{hmmer_results}
\title{get the dataframe of PFAM results from a `buscar` search object from `buscar()`}
\usage{
hmmer_results(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of PFAM results from a `buscar` search object from `buscar()`
}
