% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{keep_phobius_hits}
\alias{keep_phobius_hits}
\title{Extract proteins with Phobius hits from input file and return a tmpfile fasta
of those proteins for subsequent searches}
\usage{
keep_phobius_hits(df, file_path)
}
\arguments{
\item{df}{dataframe of Phobius hits}

\item{file_path}{path to protein fasta file}
}
\value{
tmpfile path
}
\description{
Extract proteins with Phobius hits from input file and return a tmpfile fasta
of those proteins for subsequent searches
}
