% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{lrr_rk}
\alias{lrr_rk}
\title{get the dataframe of lrr_rk results from a `busco` search object from `buscar()`}
\usage{
lrr_rk(b)
}
\arguments{
\item{b}{busco search object returned from `busco`}
}
\value{
dataframe
}
\description{
get the dataframe of lrr_rk results from a `busco` search object from `buscar()`
}
