% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{mesa}
\alias{mesa}
\title{make a table of the number of each type of receptor from an internet search}
\usage{
mesa(b)
}
\arguments{
\item{b}{an object of class `busco`, usually the search results from `buscar()`}
}
\value{
knitr::kable object
}
\description{
make a table of the number of each type of receptor from an internet search
}
