% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_raw_blast}
\alias{parse_raw_blast}
\title{parse blast result file}
\usage{
parse_raw_blast(blast_raw)
}
\arguments{
\item{blast_raw}{blast result file (`-outfmt 6`)}
}
\value{
dataframe
}
\description{
parse blast result file
}
