% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_raw_deeptmhmm}
\alias{parse_raw_deeptmhmm}
\title{process raw deeptmhmm result into basic dataframe}
\usage{
parse_raw_deeptmhmm(dtm_raw)
}
\arguments{
\item{dtm}{complete deeptmhmm result}
}
\value{
dataframe
}
\description{
process raw deeptmhmm result into basic dataframe
}
