% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_raw_ecto}
\alias{parse_raw_ecto}
\title{parse blast result file}
\usage{
parse_raw_ecto(ecto_raw)
}
\arguments{
\item{ecto_raw}{blast result file (`-outfmt 7`)}
}
\value{
dataframe
}
\description{
parse blast result file
}
