% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_raw_hmmer}
\alias{parse_raw_hmmer}
\title{process raw hmmer result into dataframe}
\usage{
parse_raw_hmmer(hmmer_raw, pfam_eval)
}
\arguments{
\item{hmmer_raw}{complete hmmer result}
}
\value{
dataframe
}
\description{
process raw hmmer result into dataframe
}
