% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{pfam_results}
\alias{pfam_results}
\title{get the dataframe of ALL PFAM results from a `buscar` search object from `buscar()`}
\usage{
pfam_results(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of ALL PFAM results from a `buscar` search object from `buscar()`
}
