% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{phobius_results}
\alias{phobius_results}
\title{get the dataframe of Phobius results from a `busco` search object from `buscar()`}
\usage{
phobius_results(b)
}
\arguments{
\item{b}{busco search object returned from `busco`}
}
\value{
dataframe
}
\description{
get the dataframe of Phobius results from a `busco` search object from `buscar()`
}
