% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{process_deeptmhmm}
\alias{process_deeptmhmm}
\title{get filtered deeptmhmm hits}
\usage{
process_deeptmhmm(dtm)
}
\arguments{
\item{dtm}{parsed deeptmhmm result from `process_raw_deeptmhmm`}
}
\value{
character vector
}
\description{
get filtered deeptmhmm hits
}
