% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phobius.R
\name{process_phobius}
\alias{process_phobius}
\title{filter and reformat raw phobius result}
\usage{
process_phobius(pr)
}
\arguments{
\item{pr}{phobius result dataframe from `get_phobius()`}
}
\value{
dataframe
}
\description{
keeps only Phobius hits with one TM and returns dataframe
}
