% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfam.R
\name{retrieve_pfam}
\alias{retrieve_pfam}
\title{check the PFAM server for an update on job progress}
\usage{
retrieve_pfam(busc)
}
\arguments{
\item{busc}{buscador object}
}
\value{
buscador object with updated pfam_progress slot or buscador object with
complete pfam slot
}
\description{
looks at the buscador$pfam_progress object and checks progress of ID on PFAM.
If all IDs are complete and retrieved, then the individual dataframes are combined
and put into buscador$pfam and given class "complete". If not the
buscador$pfam_progress is updated as far as possible but the class remains
"submission"
}
