% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{seq_to_df}
\alias{seq_to_df}
\title{load the protein fasta file in to a dataframe}
\usage{
seq_to_df(b)
}
\arguments{
\item{b}{`busco` search object returned from `buscar()`}
}
\description{
load the protein fasta file in to a dataframe
}
