% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfam.R
\name{submit_pfam}
\alias{submit_pfam}
\title{Query PFAM web server.}
\source{
\url{https://www.ebi.ac.uk/Tools/common/tools/help/index.html}
}
\usage{
submit_pfam(
  fasta_str,
  email,
  progress = FALSE,
  eval = 20,
  maxchecktime = 120,
  wait = 10
)
}
\arguments{
\item{email}{Character, a valid user email address, needed for PFAMscan use}

\item{progress}{Boolean, whether to report progress bar, at default set to
FALSE.}

\item{eval}{PFAM hit eval cutoff}

\item{file}{A character vector of FASTA protein amino acid sequences}
}
\value{
The buscador$pfam object populated with submission query ids and marked with class 'submission'
or if search completed in time a data frame with class 'complete':
}
\description{
Query PFAM web server.
}
\references{
Madeira F, Park YM, Lee J, et al. The EMBL-EBI search and sequence analysis
tools APIs in 2019. Nucleic Acids Research. 2019 Jul;47(W1):W636-W641.
\doi{10.1093/nar/gkz268}.
}
