% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{tostrvec}
\alias{tostrvec}
\title{convert aastrset to a fasta character vector}
\usage{
tostrvec(aastrset)
}
\arguments{
\item{aastrset}{Biostrings::aastrset object}
}
\value{
character vector with fasta strings
}
\description{
convert aastrset to a fasta character vector
}
