% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{totempfile}
\alias{totempfile}
\title{convert aastrset to a tempfile on disk}
\usage{
totempfile(aastrset)
}
\arguments{
\item{aastrset}{Biostrings::aastrset object}
}
\value{
filename
}
\description{
convert aastrset to a tempfile on disk
}
