% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{buscar}
\alias{buscar}
\title{perform internet searches and classify given proteins into receptor types}
\usage{
buscar(
  protein_file = NULL,
  restart_file = NULL,
  progress = FALSE,
  email = NULL,
  pfam_eval_cutoff = 20,
  blast_eval_cutoff = 1e-06,
  wait = 5,
  maxchecktime = 120
)
}
\arguments{
\item{protein_file}{path to protein fasta file (ignored if restart file has some searches). Do not use if you want to restart}

\item{restart_file}{file to use as restart file, will perform the search actions and automatically save each step as you go. if file exists will be appended with new search results. search actions already performed and saved in the file will not be replaced}

\item{progress}{show a progress bar as job runs (not recommended for non interactive sessions)}

\item{email}{a valid email address}

\item{pfam_eval_cutoff}{exclude pfam hits with eval over this value (20)}

\item{blast_eval_cutoff}{exclude blast hits with eval over this value (1e-6)}

\item{wait}{seconds to wait between PFAM checks (5)}

\item{maxchecktime}{seconds to stop checking PFAM and abort (120)}
}
\value{
list of tidy dataframes of each type of receptor and its Phobius, PFAM and BLAST hits
}
\description{
runs the internet search using the Phobius server, the PFAM API and local BLAST. Then collates the results
and classifies proteins into receptor types as defined in `definiciones()`
}
