% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{buscar}
\alias{buscar}
\title{perform internet searches and classify given proteins into receptor types}
\usage{
buscar(
  file_path,
  progress = FALSE,
  email = NULL,
  pfam_eval_cutoff = 20,
  blast_eval_cutoff = 1e-06,
  wait = 5,
  maxchecktime = 120
)
}
\arguments{
\item{file_path}{path to protein fasta file}

\item{progress}{show a progress bar as job runs (not recommended for non interactive sessions)}

\item{email}{a valid email address}

\item{pfam_eval_cutoff}{exclude pfam hits with eval over this value (20)}

\item{blast_eval_cutoff}{exclude blast hits with eval over this value (1e-6)}

\item{wait}{seconds to wait between PFAM checks (5)}

\item{maxchecktime}{seconds to stop checking PFAM and abort (120)}
}
\value{
list of tidy dataframes of each type of receptor and its Phobius, PFAM and BLAST hits, class 'busco'
}
\description{
runs the internet search using the Phobius server, the PFAM API and local BLAST. Then collates the results
and classifies proteins into receptor types as defined in `definiciones()`
}
