% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{dibujar}
\alias{dibujar}
\title{draw each found proteins of a given receptor type.}
\usage{
dibujar(b, which = "lrr_rp", label_domains = FALSE)
}
\arguments{
\item{b}{`busco` search object returned from `buscar()`}

\item{which}{the receptor type to return, one of 'lrr_rp', 'lrr_rk', 'non_lrr_rp', 'non_lrr_rk', 'lrr_rp_ecto'}

\item{label_domains}{write a label on the domain}
}
\value{
ggplot2
}
\description{
draw each found proteins of a given receptor type.
}
