% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{do_searches}
\alias{do_searches}
\title{run searches over the internet and local BLAST.}
\usage{
do_searches(
  file_path,
  progress = FALSE,
  email = NULL,
  pfam_eval = 20,
  blast_eval = 1e-06
)
}
\arguments{
\item{file_path}{path to protein fasta file}

\item{progress}{show a progress bar as job runs (not recommended for non interactive sessions)}

\item{email}{a valid email address}

\item{pfam_eval}{exclude pfam hits with eval over this value (20)}

\item{blast_eval}{exclude blast hits with eval over this value (1e-6)}
}
\value{
list of dataframes of search results
}
\description{
Runs searches for proteins at Phobius, PFAM and local BLAST.
Reformat returned results from searches and apply basic quality filters
}
