% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{lrr_rp_rk_with_ecto}
\alias{lrr_rp_rk_with_ecto}
\title{get the dataframe of lrr_rp_rk_with_ecto results from a `buscar` search object from `buscar()`}
\usage{
lrr_rp_rk_with_ecto(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of lrr_rp_rk_with_ecto results from a `buscar` search object from `buscar()`
}
