% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{non_lrr_rk}
\alias{non_lrr_rk}
\title{get the dataframe of non_lrr_rk results from a `buscar` search object from `buscar()`}
\usage{
non_lrr_rk(b)
}
\arguments{
\item{b}{buscar search object returned from `buscar`}
}
\value{
dataframe
}
\description{
get the dataframe of non_lrr_rk results from a `buscar` search object from `buscar()`
}
