% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{pfam_results}
\alias{pfam_results}
\title{get the dataframe of PFAM results from a `busco` search object from `buscar()`}
\usage{
pfam_results(b)
}
\arguments{
\item{b}{busco search object returned from `busco`}
}
\value{
dataframe
}
\description{
get the dataframe of PFAM results from a `busco` search object from `buscar()`
}
