% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfam.R
\name{process_pfam}
\alias{process_pfam}
\title{process raw pfam result into dataframe}
\usage{
process_pfam(pr, pfam_eval)
}
\arguments{
\item{pr}{complete PFAM result}
}
\value{
dataframe
}
\description{
process raw pfam result into dataframe
}
