% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscador.R
\name{seq_to_df}
\alias{seq_to_df}
\title{convert the aastrset to a dataframe}
\usage{
seq_to_df(b)
}
\arguments{
\item{b}{`buscador` search object returned from `buscar()`}
}
\description{
convert the aastrset to a dataframe
}
